FIX: wrong type for --price option
git-svn-id: file:///data/svn/enum/server/trunk@10486 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ src/cli_admin/invoiceclient.cc
@@ -386,7 +386,7 @@ InvoiceClient::credit()
     */
     if (m_conf.hasOpt(INVOICE_PRICE_NAME)) {
         // invoice->setPrice(Database::Money(
-        price = 100 * m_conf.get<long>(INVOICE_PRICE_NAME);
+        price = 100 * m_conf.get<unsigned int>(INVOICE_PRICE_NAME);
         priceFilled = true;
     }
